Base logic implemented: person/movie/tv adding/deleting from diff collections, nested rv's added to detail screens, + some refactoring

+++ data/src/main/java/com/myst3ry/data/remote/pojo/tvshow/detail/LastEpisodeToAir.java
@@ -20,7 +20,7 @@ public final class LastEpisodeToAir {
     private int mShowId;
 
     @SerializedName("vote_average")
-    private int mVoteAverage;
+    private double mVoteAverage;
 
     @SerializedName("name")
     private String mName;
@@ -54,7 +54,7 @@ public final class LastEpisodeToAir {
         return mShowId;
     }
 
-    public int getVoteAverage() {
+    public double getVoteAverage() {
         return mVoteAverage;
     }
